+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/SpotlessExtensionModern.java
@@ -23,6 +23,8 @@ import org.gradle.api.tasks.TaskContainer;
 import org.gradle.api.tasks.TaskProvider;
 
 public class SpotlessExtensionModern extends SpotlessExtensionBase {
+	private final TaskProvider<RegisterDependenciesTask> registerDependenciesTask;
+
 	public SpotlessExtensionModern(Project project) {
 		super(project);
 		rootCheckTask = project.getTasks().register(EXTENSION + CHECK, task -> {
@@ -37,6 +39,13 @@ public class SpotlessExtensionModern extends SpotlessExtensionBase {
 			task.setGroup(TASK_GROUP); // no description on purpose
 		});
 
+		TaskContainer rootProjectTasks = project.getRootProject().getTasks();
+		if (!rootProjectTasks.getNames().contains(RegisterDependenciesTask.TASK_NAME)) {
+			this.registerDependenciesTask = rootProjectTasks.register(RegisterDependenciesTask.TASK_NAME, RegisterDependenciesTask.class, RegisterDependenciesTask::setup);
+		} else {
+			this.registerDependenciesTask = rootProjectTasks.named(RegisterDependenciesTask.TASK_NAME, RegisterDependenciesTask.class);
+		}
+
 		project.afterEvaluate(unused -> {
 			if (enforceCheck) {
 				project.getTasks().named(JavaBasePlugin.CHECK_TASK_NAME)
@@ -46,6 +55,11 @@ public class SpotlessExtensionModern extends SpotlessExtensionBase {
 	}
 
 	final TaskProvider<?> rootCheckTask, rootApplyTask, rootDiagnoseTask;
+
+	@Override
+	RegisterDependenciesTask getRegisterDependenciesTask() {
+		return registerDependenciesTask.get();
+	}
 
 	@SuppressWarnings("unchecked")
 	@Override