+++ testlib/src/main/java/com/diffplug/spotless/StepHarnessWithFile.java
@@ -23,67 +23,77 @@ import org.assertj.core.api.AbstractThrowableAssert;
 import org.assertj.core.api.Assertions;
 import org.junit.Assert;
 
-/** An api for testing a `FormatterStep` that doesn't depend on the File path. DO NOT ADD FILE SUPPORT TO THIS, use {@link StepHarnessWithFile} if you need that. */
-public class StepHarness implements AutoCloseable {
+/** An api for testing a `FormatterStep` that depends on the File path. */
+public class StepHarnessWithFile implements AutoCloseable {
 	private final FormatterFunc formatter;
 
-	private StepHarness(FormatterFunc formatter) {
+	private StepHarnessWithFile(FormatterFunc formatter) {
 		this.formatter = Objects.requireNonNull(formatter);
 	}
 
-	/** Creates a harness for testing steps which don't depend on the file. */
-	public static StepHarness forStep(FormatterStep step) {
+	/** Creates a harness for testing steps which do depend on the file. */
+	public static StepHarnessWithFile forStep(FormatterStep step) {
 		// We don't care if an individual FormatterStep is misbehaving on line-endings, because
 		// Formatter fixes that.  No reason to care in tests either.  It's likely to pop up when
 		// running tests on Windows from time-to-time
-		return new StepHarness(FormatterFunc.Closeable.of(
+		return new StepHarnessWithFile(FormatterFunc.Closeable.of(
 				() -> {
 					if (step instanceof FormatterStepImpl.Standard) {
 						((FormatterStepImpl.Standard<?>) step).cleanupFormatterFunc();
 					}
 				},
-				input -> LineEnding.toUnix(step.format(input, new File("")))));
+				new FormatterFunc() {
+					@Override
+					public String apply(String input) throws Exception {
+						return apply(input, new File(""));
+					}
+
+					@Override
+					public String apply(String input, File source) throws Exception {
+						return LineEnding.toUnix(step.format(input, source));
+					}
+				}));
 	}
 
-	/** Creates a harness for testing a formatter whose steps don't depend on the file. */
-	public static StepHarness forFormatter(Formatter formatter) {
-		return new StepHarness(FormatterFunc.Closeable.of(
+	/** Creates a harness for testing a formatter whose steps do depend on the file. */
+	public static StepHarnessWithFile forFormatter(Formatter formatter) {
+		return new StepHarnessWithFile(FormatterFunc.Closeable.of(
 				formatter::close,
 				input -> formatter.compute(input, new File(""))));
 	}
 
 	/** Asserts that the given element is transformed as expected, and that the result is idempotent. */
-	public StepHarness test(String before, String after) throws Exception {
-		String actual = formatter.apply(before);
+	public StepHarnessWithFile test(File file, String before, String after) throws Exception {
+		String actual = formatter.apply(before, file);
 		Assert.assertEquals("Step application failed", after, actual);
-		return testUnaffected(after);
+		return testUnaffected(file, after);
 	}
 
 	/** Asserts that the given element is idempotent w.r.t the step under test. */
-	public StepHarness testUnaffected(String idempotentElement) throws Exception {
-		String actual = formatter.apply(idempotentElement);
+	public StepHarnessWithFile testUnaffected(File file, String idempotentElement) throws Exception {
+		String actual = formatter.apply(idempotentElement, file);
 		Assert.assertEquals("Step is not idempotent", idempotentElement, actual);
 		return this;
 	}
 
 	/** Asserts that the given elements in  the resources directory are transformed as expected. */
-	public StepHarness testResource(String resourceBefore, String resourceAfter) throws Exception {
+	public StepHarnessWithFile testResource(File file, String resourceBefore, String resourceAfter) throws Exception {
 		String before = ResourceHarness.getTestResource(resourceBefore);
 		String after = ResourceHarness.getTestResource(resourceAfter);
-		return test(before, after);
+		return test(file, before, after);
 	}
 
 	/** Asserts that the given elements in the resources directory are transformed as expected. */
-	public StepHarness testResourceUnaffected(String resourceIdempotent) throws Exception {
+	public StepHarnessWithFile testResourceUnaffected(File file, String resourceIdempotent) throws Exception {
 		String idempotentElement = ResourceHarness.getTestResource(resourceIdempotent);
-		return testUnaffected(idempotentElement);
+		return testUnaffected(file, idempotentElement);
 	}
 
 	/** Asserts that the given elements in the resources directory are transformed as expected. */
-	public StepHarness testException(String resourceBefore, Consumer<AbstractThrowableAssert<?, ? extends Throwable>> exceptionAssertion) throws Exception {
+	public StepHarnessWithFile testException(File file, String resourceBefore, Consumer<AbstractThrowableAssert<?, ? extends Throwable>> exceptionAssertion) throws Exception {
 		String before = ResourceHarness.getTestResource(resourceBefore);
 		try {
-			formatter.apply(before);
+			formatter.apply(before, file);
 			Assert.fail();
 		} catch (Throwable t) {
 			AbstractThrowableAssert<?, ? extends Throwable> abstractAssert = Assertions.assertThat(t);